Moved ListItemLabel to its own file
